requirements.txt
selenium~=4.43.0
pytest~=9.0.3
allure-python-commons~=2.15.3